+++ packages/adapters/src/index.ts
@@ -1,1 +1,6 @@
 export * from "./sqlite.js";
+export * from "./matrix.js";
+export * from "./managed-postgres.js";
+export * from "./deployment-plan.js";
+export * from "./terraform.js";
+export * from "./self-hosted.js";